requirements.txt
# FastAPI 核心依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库相关
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1

# 其他工具
python-dotenv>=1.0.0
python-multipart>=0.0.6
